TFC/requirements.txt
Django==4.1.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
markdown==3.4.1
geopy==2.3.0
Pillow==9.3.0
django-cors-headers==3.13.0